fix: refactor dto
